third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/pow.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] pow float32 1D base tensor and 1D integer exponent tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test pow float32 actual NaN should be close enough to expected 6.316321332633379e-8 by ULP distance: expected a number less than or equal to 32n but got 1278761896n
[FAIL] pow float32 2D base tensor and 2D integer exponent tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test pow float32 actual NaN should be close enough to expected 6.316321332633379e-8 by ULP distance: expected a number less than or equal to 32n but got 1278761896n
[FAIL] pow float32 3D base tensor and 3D integer exponent tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test pow float32 actual NaN should be close enough to expected 6.316321332633379e-8 by ULP distance: expected a number less than or equal to 32n but got 1278761896n
[FAIL] pow float32 4D base tensor and 4D integer exponent tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test pow float32 actual NaN should be close enough to expected 6.316321332633379e-8 by ULP distance: expected a number less than or equal to 32n but got 1278761896n
[FAIL] pow float32 5D base tensor and 5D integer exponent tensor
  promise_test: Unhandled rejection with value: object "TypeError: Failed to execute 'pow' on 'MLGraphBuilder': Unsupported rank 5 for argument a (must be at most 4)."
[FAIL] pow (square) float32 4D base tensor and broadcastable 0D integer exponent scalar
  assert_less_than_equal: assert_array_approx_equals_ulp: test pow float32 actual NaN should be close enough to expected 0.00398249039426446 by ULP distance: expected a number less than or equal to 32n but got 1144881267n
[FAIL] pow (sqrt) float32 4D positive base tensor and broadcastable 0D integer exponent scalar
  assert_less_than_equal: assert_array_approx_equals_ulp: test pow float32 actual 4.303055286407471 should be close enough to expected 1.5685393810272217 by ULP distance: expected a number less than or equal to 32n but got 12643515n
[FAIL] pow float32 4D base tensor and broadcastable 2D integer exponent tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test pow float32 actual NaN should be close enough to expected 998220038144 by ULP distance: expected a number less than or equal to 32n but got 743937395n
[FAIL] pow float32 4D base tensor and broadcastable 3D integer exponent tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test pow float32 actual NaN should be close enough to expected -999109.625 by ULP distance: expected a number less than or equal to 32n but got 3375623258n
[FAIL] pow float16 constant 1D base tensor and 1D integer exponent tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test pow float16 actual -2 should be close enough to expected 17.84375 by ULP distance: expected a number less than or equal to 2 but got 29578
[FAIL] pow float16 1D base tensor and 1D integer exponent tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test pow float16 actual 0 should be close enough to expected 17.84375 by ULP distance: expected a number less than or equal to 2 but got 19574
[FAIL] pow float16 2D base tensor and 2D integer exponent tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test pow float16 actual 0 should be close enough to expected 17.84375 by ULP distance: expected a number less than or equal to 2 but got 19574
[FAIL] pow float16 3D base tensor and 3D integer exponent tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test pow float16 actual 0 should be close enough to expected 17.84375 by ULP distance: expected a number less than or equal to 2 but got 19574
[FAIL] pow float16 4D base tensor and 4D integer exponent tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test pow float16 actual 0 should be close enough to expected 17.84375 by ULP distance: expected a number less than or equal to 2 but got 19574
[FAIL] pow float16 5D base tensor and 5D integer exponent tensor
  promise_test: Unhandled rejection with value: object "TypeError: Failed to execute 'pow' on 'MLGraphBuilder': Unsupported rank 5 for argument a (must be at most 4)."
[FAIL] pow (square) float16 4D base tensor and broadcastable 0D integer exponent scalar
  assert_less_than_equal: assert_array_approx_equals_ulp: test pow float16 actual 0 should be close enough to expected 318.5 by ULP distance: expected a number less than or equal to 2 but got 23802
[FAIL] pow (sqrt) float16 4D positive base tensor and broadcastable 0D integer exponent scalar
  assert_less_than_equal: assert_array_approx_equals_ulp: test pow float16 actual 0 should be close enough to expected 1.19140625 by ULP distance: expected a number less than or equal to 2 but got 15556
[FAIL] pow float16 4D base tensor and broadcastable 2D integer exponent tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test pow float16 actual 0 should be close enough to expected 5680 by ULP distance: expected a number less than or equal to 2 but got 28044
[FAIL] pow float16 4D base tensor and broadcastable 3D integer exponent tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test pow float16 actual 0 should be close enough to expected 0.0560302734375 by ULP distance: expected a number less than or equal to 2 but got 11052
Harness: the test ran to completion.
